pyproject.toml
[tool.poetry]
name = "ramws"
version = "0.0.0"
description = "Gestión de calidad del aire de forma inteligente."
authors = [
  "Ariel Martínez <ariel@martinezs.dev>",
]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
Django = "^3.2.5"
dj-database-url = "^0.5.0"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"
whitenoise = "^5.2.0"
django-extensions = "^3.1.3"
psycopg2-binary = "^2.9.1"
gunicorn = "^20.1.0"
djangorestframework-simplejwt = "^4.7.2"
boto3 = "^1.18.2"
sentry-sdk = "^1.3.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
isort = "^5.9.2"
pydotplus = "^2.0.2"

[tool.black]
line_length = 90
target_version = ["py36", "py37", "py38"]
exclude = '/(\.venv|migrations)/'

[tool.isort]
profile = "black"
multi_line_output = 3
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_django = "django"
known_first_party = ["accounts", "docs", "core", "utils"]

[build-system]
requires = ["poetry>=1.1.7"]
build-backend = "poetry.masonry.api"
